requirements.txt
selenium
webdriver-manager
beautifulsoup4
lxml
python-docx
azure-ai-textanalytics
python-dotenv